@@ -1,10 +1,9 @@
-import Vue from 'vue'
+import { createApp, h } from 'vue';
 import App from './App.vue'
 import router from './router'
+import store from './store'
 
-Vue.config.productionTip = false
-
-new Vue({
-  router,
-  render: h => h(App),
-}).$mount('#app')
+const app = createApp(App)
+app.use(store)
+app.use(router)
+app.mount('#app')
